+++ CHANGELOG.md
@@ -4,6 +4,15 @@
 
 ## [Unreleased]
 
+## [1.1.0] - 2023-07-04
+
+### Added
+- Assisted code completion with @Table annotation.
+- Support ORM's table method.
+
+### Fixed
+- Fix SQL to Struct error.
+
 ## [1.0.2] - 2023-07-04
 
 ### Fixed
@@ -25,7 +34,8 @@
 - ORM Code Completion
 - SQL to Struct
 
-[Unreleased]: https://github.com/maiqingqiang/go-orm-helper/compare/v1.0.2...HEAD
+[Unreleased]: https://github.com/maiqingqiang/go-orm-helper/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/maiqingqiang/go-orm-helper/compare/v1.0.2...v1.1.0
 [1.0.2]: https://github.com/maiqingqiang/go-orm-helper/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/maiqingqiang/go-orm-helper/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/maiqingqiang/go-orm-helper/tree/v1.0.0